Week 6
I created a text-adventure game called "Inner City,"using HTML, CSS, and Javascript. 
It is about mantaining the mental, emotional, and physical well-being of one of two characters
living in the city. The inner city is not just where they live, it is the life they lead
and it is the player's job to ensure they live a good one, OR end up suffering from a meltdown, 
and losing the game.